# backend-server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.StorageBucket:
    properties:
      accessKey:
        type: string
      bucket:
        type: string
      cdnDomain:
        type: string
      createdAt:
        type: string
      description:
        type: string
      driver:
        type: string
      endpoint:
        type: string
      id:
        type: integer
      isDefault:
        type: boolean
      name:
        type: string
      pathPrefix:
        type: string
      purpose:
        type: string
      region:
        type: string
      secretKey:
        type: string
      status:
        type: integer
      updatedAt:
        type: string
      useSsl:
        type: boolean
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      error: {}
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: 后端管理服务 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Backend Server API
  version: "1.0"
paths:
...
  /system/storage-buckets:
    get:
      description: 获取所有存储桶配置
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.StorageBucket'
                  type: array
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取存储桶列表
      tags:
      - 存储桶
    post:
      consumes:
      - application/json
      description: 创建新的存储桶配置
      parameters:
      - description: 存储桶配置
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.StorageBucket'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.StorageBucket'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建存储桶
      tags:
      - 存储桶
  /system/storage-buckets/{id}:
    delete:
      description: 删除指定存储桶配置
      parameters:
      - description: 存储桶 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除存储桶
      tags:
      - 存储桶
    get:
      description: 根据 ID 获取存储桶配置详情
      parameters:
      - description: 存储桶 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.StorageBucket'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取存储桶详情
      tags:
      - 存储桶
    put:
      consumes:
      - application/json
      description: 更新指定存储桶配置
      parameters:
      - description: 存储桶 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 存储桶配置
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.StorageBucket'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.StorageBucket'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新存储桶
      tags:
      - 存储桶
  /system/storage-buckets/{id}/default:
    put:
      description: 将指定存储桶设为默认
      parameters:
      - description: 存储桶 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 设置默认存储桶
      tags:
      - 存储桶
  /system/storage-buckets/{id}/test:
    post:
      description: 测试指定存储桶连接是否可用
      parameters:
      - description: 存储桶 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 测试存储桶连接
      tags:
      - 存储桶
  /system/storage-buckets/default:
    get:
      description: 获取当前默认存储桶配置
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.StorageBucket'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取默认存储桶
      tags:
      - 存储桶
  /system/storage-buckets/driver/{driver}:
    get:
      description: 根据存储驱动获取存储桶列表
      parameters:
      - description: 存储驱动
        in: path
        name: driver
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.StorageBucket'
                  type: array
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 按驱动获取存储桶
      tags:
      - 存储桶
  /system/storage-buckets/enabled-drivers:
    get:
      description: 获取当前已启用的存储驱动列表
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取已启用存储驱动
      tags:
      - 存储桶
  /system/storage-buckets/purpose/{purpose}:
    get:
      description: 根据用途获取存储桶列表
      parameters:
      - description: 用途
        in: path
        name: purpose
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.StorageBucket'
                  type: array
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 按用途获取存储桶
      tags:
      - 存储桶
  /system/storage-buckets/test-by-driver:
    post:
      consumes:
      - application/json
      description: 根据驱动和桶名测试连接，无需保存配置
      parameters:
      - description: 测试参数
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 按驱动测试连接
      tags:
      - 存储桶
...
securityDefinitions:
  BearerAuth:
    description: JWT Token（格式：Bearer {token}）
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"